+++ src/app/(dashboard)/educators/_components/dashboard/DashboardTaskList.tsx
@@ -6,7 +6,7 @@ import { ChevronRight } from "lucide-react";
 import { TeacherProfileAvatar } from "@/components/common/avatar/TeacherProfileAvatar";
 import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
-import { useAuthContext } from "@/providers/AuthProvider";
+import { useAuthContext } from "@/shared/common/providers/AuthProvider";
 import { DashboardTask } from "@/types/dashboard";
 
 type DashboardTaskListProps = {
