+++ src/app/(dashboard)/educators/_hooks/useEducatorsDashboardPage.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 
 import { dashboardKeys } from "@/constants/query-keys";
-import { useAuthContext } from "@/providers/AuthProvider";
+import { useAuthContext } from "@/shared/common/providers/AuthProvider";
 import {
   mapMgmtAssistantOrdersToTasks,
   mapMgmtDashboardToClinics,
