+++ src/app/(dashboard)/educators/communication/[communicationId]/page.tsx
@@ -17,7 +17,7 @@ import {
 import { useDownloadMaterial } from "@/hooks/useMaterials";
 import { CommonPostAttachment } from "@/types/communication/commonPost";
 import { useDialogAlert } from "@/hooks/useDialogAlert";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import { CheckModal } from "@/components/common/modals/CheckModal";
 import {
   GetInstructorPostDetailResponse,
