+++ src/app/(dashboard)/educators/exams/create/_hooks/useExamCreateForm.tsx
@@ -4,7 +4,7 @@ import { useRouter } from "next/navigation";
 import { useFieldArray, useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import { CheckModal } from "@/components/common/modals/CheckModal";
 import {
   examFormSchema,
