+++ src/app/(dashboard)/educators/materials/[materialsId]/page.tsx
@@ -13,7 +13,7 @@ import {
   VideoFormData,
   RequestFormData,
 } from "@/types/materials.type";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import { CheckModal } from "@/components/common/modals/CheckModal";
 import {
   useDownloadMaterial,
