+++ src/app/(dashboard)/educators/profile/_components/modal/PhoneChangeModal.tsx
@@ -12,7 +12,7 @@ import {
 } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
 import { InputForm } from "@/components/common/input/InputForm";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import { formatPhoneNumber } from "@/utils/phone";
 import {
   phoneChangeSchema,
