+++ src/app/(dashboard)/educators/profile/_components/modal/SettingsSecurityModal.tsx
@@ -13,8 +13,8 @@ import {
 } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
 import { InputForm } from "@/components/common/input/InputForm";
-import { useModal } from "@/providers/ModalProvider";
-import { useAuthContext } from "@/providers/AuthProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
+import { useAuthContext } from "@/shared/common/providers/AuthProvider";
 import { CheckModal } from "@/components/common/modals/CheckModal";
 import { verifyEmailAPI } from "@/services/auth.service";
 import { changeMyPasswordAPI } from "@/services/profile.service";
