+++ src/app/(dashboard)/educators/schedules/_hooks/useScheduleEditor.ts
@@ -5,7 +5,7 @@ import { createElement } from "react";
 import { format } from "date-fns";
 
 import { CheckModal } from "@/components/common/modals/CheckModal";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import {
   mapScheduleFormToPayload,
   mapScheduleFormToUpdatePayload,
