+++ src/app/(dashboard)/educators/schedules/_hooks/useScheduleEvents.ts
@@ -11,7 +11,7 @@ import type { View } from "react-big-calendar";
 import { isSameDay } from "date-fns";
 
 import { CheckModal } from "@/components/common/modals/CheckModal";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import {
   buildSchedulesRangeQuery,
   mapScheduleApiToCalendarEvent,
