+++ src/app/(dashboard)/educators/students/[studentId]/_components/detail-modal/AttendanceDetailModal.tsx
@@ -12,7 +12,7 @@ import {
   DialogFooter,
 } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import {
   GetEnrollmentDetail,
   AttendanceList,
