+++ src/app/(dashboard)/educators/students/[studentId]/_components/detail-modal/AttendanceRegisterModal.tsx
@@ -8,7 +8,7 @@ import {
   DialogDescription,
   DialogTitle,
 } from "@/components/ui/dialog";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import {
   AttendanceRegisterFormData,
   AttendanceStatus,
