+++ src/app/(dashboard)/educators/students/[studentId]/_components/detail-modal/EditProfileModal.tsx
@@ -11,7 +11,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import { Label } from "@/components/ui/label";
 import { Button } from "@/components/ui/button";
 import {
