+++ src/app/(dashboard)/educators/students/_components/action/StudentActions.tsx
@@ -1,7 +1,7 @@
 import { ArrowLeftRight, CheckCircle2, Send, UserPlus } from "lucide-react";
 
 import { Button } from "@/components/ui/button";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import { LectureStatus } from "@/types/students.type";
 
 import { StudentChangeModal } from "../modal/ClassChangeModal";
