+++ src/app/(dashboard)/educators/students/_components/modal/ClassChangeModal.tsx
@@ -13,7 +13,7 @@ import {
   DialogTitle,
 } from "@/components/ui/dialog";
 import { Textarea } from "@/components/ui/textarea";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import { useStudentSelectionStore } from "@/stores/studentsList.store";
 import SelectBtn from "@/components/common/button/SelectBtn";
 import { useLecturesList, useMigrateStudents } from "@/hooks/useEnrollment";
