+++ src/app/(dashboard)/educators/students/_components/modal/StudentCreateModal.tsx
@@ -14,7 +14,7 @@ import {
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import { StudentCreateFormData } from "@/types/students.type";
 import {
   getCreateStudentFormDefaults,
