+++ src/app/(dashboard)/learners/lectures/_components/AttendanceDetailModal.tsx
@@ -7,7 +7,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import { AttendanceList, AttendanceStatus } from "@/types/students.type";
 import { formatYMDFromISO } from "@/utils/date";
 import { formatTextWithLineBreaks } from "@/utils/formatTextWithLineBreaks";
