+++ src/app/(dashboard)/learners/profile/_components/modal/LearnersProfileEditModal.tsx
@@ -12,7 +12,7 @@ import {
   DialogTitle,
 } from "@/components/ui/dialog";
 import { InputForm } from "@/components/common/input/InputForm";
-import { useModal } from "@/providers/ModalProvider";
+import { useModal } from "@/shared/common/providers/ModalProvider";
 import { formatPhoneNumber } from "@/utils/phone";
 import {
   learnersProfileUpdateSchema,
