+++ src/components/common/sidebar/AppSidebar.tsx
@@ -26,9 +26,9 @@ import {
   SidebarMenuItem,
   SidebarFooter,
 } from "@/components/ui/sidebar";
-import { useAuthContext } from "@/providers/AuthProvider";
+import { useAuthContext } from "@/shared/common/providers/AuthProvider";
 import { Role } from "@/types/auth.type";
-import { API_URL_TYPE, useAuth } from "@/hooks/useAuth";
+import { API_URL_TYPE, useAuth } from "@/shared/common/hooks/useAuth";
 
 const instructorMenuItems = [
   {
