+++ src/hooks/useBreadcrumb.ts
@@ -1,6 +1,9 @@
 import { useEffect, useRef } from "react";
 
-import { useBreadcrumb, BreadcrumbItem } from "@/providers/BreadcrumbProvider";
+import {
+  useBreadcrumb,
+  BreadcrumbItem,
+} from "@/shared/common/providers/BreadcrumbProvider";
 
 export function useSetBreadcrumb(items: BreadcrumbItem[]) {
   const { setBreadcrumbs } = useBreadcrumb();
